requirements.txt
streamlit
numpy 
Pillow
opencv-python-headles 
pandas